payment-service/requirements.txt
fastapi
uvicorn[standard]
mercadopago
pydantic-settings
